dashboard/requirements.txt
streamlit==1.28.1
plotly==5.18.0
pandas==2.1.2
numpy==1.26.1
python-dotenv==1.0.0
requests==2.31.0
